Added rounding

+++ AutoTf.AdminPanel.Models/Requests/MemoryStats.cs
@@ -2,7 +2,7 @@ namespace AutoTf.AdminPanel.Models.Requests;
 
 public class MemoryStats
 {
-    public double MemoryUsageMb { get; set; }
-    public double MemoryLimitMb { get; set; }
-    public double MemoryPercentage { get; set; }
+    public float MemoryUsageMb { get; set; }
+    public float MemoryLimitMb { get; set; }
+    public float MemoryPercentage { get; set; }
 }
